Remove windows support
