+++ .gitignore
@@ -2,8 +2,6 @@
 pkg/assets/zz_generated_offsets*.go
 k0s
 k0s.code
-k0s.exe
-k0s.exe.code
 bindata*
 .*.stamp
 .tmp
