+++ pkg/backup/filesystem.go
@@ -1,6 +1,3 @@
-//go:build !windows
-// +build !windows
-
 /*
 Copyright 2022 k0s authors
