+++ pkg/component/controller/kubeletconfig_test.go
@@ -46,10 +46,8 @@ func Test_KubeletConfig(t *testing.T) {
 		t.Run("output_must_have_3_manifests", func(t *testing.T) {
 			require.Len(t, manifestYamls, 4, "Must have exactly 4 generated manifests per profile")
 			requireConfigMap(t, manifestYamls[0], "kubelet-config-default-1.24")
-			requireConfigMap(t, manifestYamls[1], "kubelet-config-default-windows-1.24")
 			requireRole(t, manifestYamls[2], []string{
 				formatProfileName("default"),
-				formatProfileName("default-windows"),
 			})
 			requireRoleBinding(t, manifestYamls[3])
 		})
@@ -77,7 +75,7 @@ func Test_KubeletConfig(t *testing.T) {
 		t.Run("final_output_must_have_manifests_for_profiles", func(t *testing.T) {
 			// check that each profile has config map, role and role binding
 			var resourceNamesForRole []string
-			for idx, profileName := range []string{"default", "default-windows", "profile_XXX", "profile_YYY"} {
+			for idx, profileName := range []string{"default", "profile_XXX", "profile_YYY"} {
 				fullName := "kubelet-config-" + profileName + "-1.24"
 				resourceNamesForRole = append(resourceNamesForRole, formatProfileName(profileName))
 				requireConfigMap(t, manifestYamls[idx], fullName)
