+++ pkg/constant/constant_posix_test.go
